added base url of application to osm config

+++ de.iteratec.hh.osm/entrypoint.sh
@@ -8,10 +8,11 @@ sed -i "s/\${TOMCAT_PASSWORD}/$TOMCAT_PASSWORD/g" /usr/local/tomcat/conf/tomcat-
 
 # prepare osm config respective set env variables
 OSM_CONFIG_TARGET_LOCATION="/root/.grails/OpenSpeedMonitor-config.groovy"
-if [ -n "$OSM_ADMIN_USER" ] && [ -n "$OSM_ADMIN_PASSWORD" ] && [ -n "$OSM_ROOT_USER" ] && [ -n "$OSM_ROOT_PASSWORD" ] && [ -n "$MYSQL_DATABASE" ] && [ -n "$MYSQL_USER" ] && [ -n "$MYSQL_PASSWORD" ]
+if [ -n "BASE_URL_OSM" ] && [ -n "$OSM_ADMIN_USER" ] && [ -n "$OSM_ADMIN_PASSWORD" ] && [ -n "$OSM_ROOT_USER" ] && [ -n "$OSM_ROOT_PASSWORD" ] && [ -n "$MYSQL_DATABASE" ] && [ -n "$MYSQL_USER" ] && [ -n "$MYSQL_PASSWORD" ]
 then
     echo "complete osm config file gets prepared"
 	cp -f /osm-templates/osm-config.groovy $OSM_CONFIG_TARGET_LOCATION
+	sed -i "s/\${BASE_URL_OSM}/$BASE_URL_OSM/g" $OSM_CONFIG_TARGET_LOCATION
 	sed -i "s/\${OSM_ADMIN_USER}/$OSM_ADMIN_USER/g" $OSM_CONFIG_TARGET_LOCATION
 	sed -i "s/\${OSM_ADMIN_PASSWORD}/$OSM_ADMIN_PASSWORD/g" $OSM_CONFIG_TARGET_LOCATION
 	sed -i "s/\${OSM_ROOT_USER}/$OSM_ROOT_USER/g" $OSM_CONFIG_TARGET_LOCATION
@@ -20,7 +21,7 @@ then
 	sed -i "s/\${MYSQL_USER}/$MYSQL_USER/g" $OSM_CONFIG_TARGET_LOCATION
 	sed -i "s/\${MYSQL_PASSWORD}/$MYSQL_PASSWORD/g" $OSM_CONFIG_TARGET_LOCATION
 	cp -f /osm-templates/setenv.sh /usr/local/tomcat/bin
-elif [ -z "$OSM_ADMIN_USER" ] && [ -z "$OSM_ADMIN_PASSWORD" ] && [ -z "$OSM_ROOT_USER" ] && [ -z "$OSM_ROOT_PASSWORD" ] && [ -z "$MYSQL_DATABASE" ] && [ -z "$MYSQL_USER" ] && [ -z "$MYSQL_PASSWORD" ] && [ ! -e $OSM_CONFIG_TARGET_LOCATION ]
+elif [ -z "BASE_URL_OSM" ] && [ -z "$OSM_ADMIN_USER" ] && [ -z "$OSM_ADMIN_PASSWORD" ] && [ -z "$OSM_ROOT_USER" ] && [ -z "$OSM_ROOT_PASSWORD" ] && [ -z "$MYSQL_DATABASE" ] && [ -z "$MYSQL_USER" ] && [ -z "$MYSQL_PASSWORD" ] && [ ! -e $OSM_CONFIG_TARGET_LOCATION ]
 then
 	echo "osm config file with some default users gets prepared"
 	cp -f /osm-templates/osm-just-users-config.groovy $OSM_CONFIG_TARGET_LOCATION
